tracing: Add collector URI to the Tracer config
Either we get the collector URI from the cluster
configuration data or we get it from the SSNTP certs.

With the former, it is up to the tracer creator to pass
it through the Tracer configuration structure to the
Tracer itself.

Since we can potentially also get it from the SSNTP
cert, the Tracer config URI is an optional field.

Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>