tracing: Queue an send spans
Spans gathered by the tracer routine are now queued and eventually
pushed to a collector through SSNTP.

At the moment we decide to empty the span queue only when its length
is getting longer than 16 spans. In the longer term we should be
able to gather SSNTP information allowing us to push this data when
control plane bandwidth is available.

Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>